Move fansite.py to modules feed and streaming client.py
