+++ src/modules/twitter/feed.py
@@ -1,78 +1,13 @@
-import asyncio
-import json
 import logging
-import os
-import re
 from typing import List, Literal
 
 import discord
 import tweepy
-from tweepy.asynchronous import AsyncStreamingClient
 
-from src.cogs.content_poster.ui import PersistentTweetView, TweetDetails
-from src.utils.config import ContentPosterConfig
-from src.utils.helper import convert_files_to_zip, download_files, get_from_dict
+from src.modules.twitter.streaming_client import TwitterStreamingClient
 
 
-class FansiteStreamingClient(AsyncStreamingClient):
-    def __init__(self, client: discord.Client):
-        # The `wait_on_rate_limit` argument prevents the streaming client from shutting off when the API rate limit is reached
-        super().__init__(bearer_token=os.getenv("TWITTER_BEARER_TOKEN"), wait_on_rate_limit=True, max_retries=5)
-        self.client = client
-        self.channel = ContentPosterConfig().get_feed_channel(self.client)
-
-    async def send_post(self, data):
-        """Sends a post with a PersistentTweetView to the feed channel."""
-        # Get all the required information from the incoming data
-        medias = get_from_dict(data, ["includes", "media"])
-        user = get_from_dict(data, ["includes", "users"])[0]
-        tweet_text = get_from_dict(data, ["data", "text"])
-        urls = [media["url"] for media in medias]
-
-        # Download the files from the URL and create a ZIP file
-        files = await download_files(urls)
-        zip_file = await convert_files_to_zip(files)
-        feed_channel_files = [*files, zip_file]  # Combines the Media and ZIP files, only posted in the feed channel
-
-        # Post the Twitter message to the feed channel
-        message = await self.channel.send(content=user["name"], files=feed_channel_files)
-
-        # The following for loop is to make it so that the Discord files are read from the first byte again after being sent as a message earlier
-        # Being sent as a message initially means the byte-file pointer is at the end
-        files = list(map(lambda f: f.fp.seek(0), files))
-        # ? Maybe use a more elegant method here instead of a for loop
-        # for f in files:
-        #     f.fp.seek(0)
-
-        # The following RegEx expression serves to obtain the Tweet URL from the caption
-        url = re.search(r"https://t.co/\S+", tweet_text)
-        tweet_details: TweetDetails = {"user": user, "url": url.group()}
-
-        # Update the message with the PersistentTweetView
-        view = PersistentTweetView(message_id=message.id, files=files, bot=self.client, tweet_details=tweet_details)
-        self.client.add_view(view=view)
-        await message.edit(view=view)
-
-        # Once the user is done with the PersistentTweetView, remove the view from the original message
-        await view.wait()
-        await message.edit(view=None)  # ? Maybe move this to the `Stop` button in the UI file
-
-    async def on_connect(self):
-        logging.info("Twitter stream has been connected successfully")
-
-    async def on_disconnect(self):
-        logging.info("Twitter stream has been disconnected successfully")
-
-    async def on_data(self, raw_data):
-        """Triggered when data is received from the stream."""
-        data = json.loads(raw_data)
-
-        # The following runs asynchronous tasks in a coroutine so that it doesn't block the main event loop
-        loop = asyncio.get_event_loop()
-        loop.create_task(self.send_post(data))
-
-
-class FansiteFeed:
+class TwitterFeed:
     # Static class variables
     rule_prefix = "(from:"
     rule_postfix = ") has:media"
@@ -82,8 +17,10 @@ class FansiteFeed:
     def __init__(self, client: discord.Client):
         # self.follow = self.get_user_ids()
         self.follow = ["1578046589578661888"]  # ! Remove this once the testing is complete
-        self.follow_change_flag = False # This flag indicates whether the follow list needs to be refreshed on Stream restart
-        self.stream: FansiteStreamingClient = None
+        self.follow_change_flag = (
+            False  # This flag indicates whether the follow list needs to be refreshed on Stream restart
+        )
+        self.stream: TwitterStreamingClient = None
 
         self.client: discord.Client = client
 
@@ -103,7 +40,7 @@ class FansiteFeed:
 
     def save_user_id(self, user_id: str, purpose: Literal["add", "remove"]):
         """Adds or removes a Twitter user ID to the `IDs.txt` file.
-        
+
         Parameters
         ----------
             * user_id: :class:`str`
@@ -130,8 +67,8 @@ class FansiteFeed:
         self, user_id: str, user_ids: List[str], rule_content: str = "", rule_contents: List[str] = []
     ):
         """
-        A recursive function that generates rule contents based on the Twitter user IDs. 
-        Only 25 `StreamRule`s with content lengths of 512 characters are allowed per `Stream`. 
+        A recursive function that generates rule contents based on the Twitter user IDs.
+        Only 25 `StreamRule`s with content lengths of 512 characters are allowed per `Stream`.
         This recursive function ensures that these requirements are met.
 
         Parameters
@@ -149,11 +86,11 @@ class FansiteFeed:
         content = f"{user_id}" if rule_content == "" else f"{self.rule_connector}{user_id}"
         temp_rule_content = f"{rule_content}{content}"
 
-        if len(user_ids) == 0: 
+        if len(user_ids) == 0:
             # If there are no more User IDs, append the rule content to the array and end the recursive loop
             rule_contents.append(f"{self.rule_prefix}{temp_rule_content}{self.rule_postfix}")
             return rule_contents
-        
+
         if len(temp_rule_content) >= self.max_rule_content_length:
             # If the new rule content exceeds 512 characters, append the old rule content and use the current user ID in the next recursion
             rule_contents.append(f"{self.rule_prefix}{rule_content}{self.rule_postfix}")
@@ -165,7 +102,7 @@ class FansiteFeed:
 
     def generate_stream_rules(self):
         """Generates a list of `StreamRule`s by using the rule contents generated by `generate_rule_contents` method."""
-        user_ids = self.follow.copy() # Copy this to prevent it from mutating the follow list
+        user_ids = self.follow.copy()  # Copy this to prevent it from mutating the follow list
         first_user_id = user_ids.pop(0)
         rule_contents = self.generate_rule_contents(first_user_id, user_ids)
         return [tweepy.StreamRule(rule_content) for rule_content in rule_contents]
@@ -177,11 +114,13 @@ class FansiteFeed:
             await self.stream.delete_rules([rule.id for rule in current_rules.data])
 
     async def start(self):
-        """Starts the `FansiteStreamingClient`."""
-        self.stream = FansiteStreamingClient(self.client) # ! Can't setup more than 1 stream concurrently if there are more than a certain amount of user IDs
+        """Starts the `TwitterStreamingClient`."""
+        self.stream = TwitterStreamingClient(
+            self.client
+        )  # ! Can't setup more than 1 stream concurrently if there are more than a certain amount of user IDs
 
         current_rules = await self.stream.get_rules()
-        if current_rules.data is None or self.follow_change_flag: 
+        if current_rules.data is None or self.follow_change_flag:
             # Only regenerate stream rules if there are no current rules or the follow list has been changed
             await self.clear_all_stream_rules()
             await self.stream.add_rules(self.generate_stream_rules())
